remove getter from extending classes, include identifier dtos to Quotation and CostEstimate objects
